Move all curses dependent code to the curses directory
