--- color/color.h
@@ -1,4 +1,0 @@
-#include <ncurses.h>
-#include "../game_cfg/colors.h"
-
-bool init_colors();
