+++ curses/color.c
