+++ curses/color.h
@@ -1,3 +1,8 @@
+#ifndef COLOR_H
+#define COLOR
+
+#include <ncurses.h>
+
 // Text colors
 #define ORIG_COLORS -1
 #define BLACK 0
@@ -8,3 +13,7 @@
 #define MAGENTA 5
 #define CYAN 6
 #define WHITE 7
+
+bool init_colors();
+
+#endif
