+++ game_cfg/game_cfg.h
@@ -1,7 +1,6 @@
 #ifndef GAME_CFG_H
 #define GAME_CFG_H
 
-#include "colors.h"
 #include "game_internals.h"
 #include "setup.h"
 #include "symbols.h"
